Assorted fixes and style updates

+++ docker/commands/get-output.sh
@@ -1,8 +1,15 @@
 #!/bin/sh
 
-uniq=$(date +%N)
 CONSOLE_LOG="/home/steam/gmodserver/garrysmod/console.log"
 TEMP_LOG="/tmp/output_to_send.$uniq.log"
+
+if [ ! -s "$CONSOLE_LOG" ]; then
+    echo "Content-Type: text/plain"
+    echo ""
+    exit 0
+fi
+
+uniq=$(date +%N)
 
 mv "$CONSOLE_LOG" "$TEMP_LOG" 2>/dev/null || true
 touch "$CONSOLE_LOG"
